requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
lxml==4.9.3
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.13.0
